+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2026-04-16]
+## [1.0.0 - 2026-04-21]
 ### Added
 - First version of the project
 - Spring Application
@@ -286,3 +286,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Jwt decoder configuration
 - Full name if the first or last name is missing
 - Only one user session at the same time
+- Update query outputs correctly